wip adding react-router
